+++ src/Caliburn.Light.WinUI/Services/IWindowManager.cs
@@ -1,4 +1,6 @@
 using Microsoft.UI.Xaml.Controls;
+using Microsoft.Windows.Storage.Pickers;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Caliburn.Light.WinUI;
@@ -38,4 +40,28 @@ public interface IWindowManager
     /// <param name="context">The context.</param>
     /// <returns>The content dialog result.</returns>
     Task<ContentDialogResult> ShowContentDialog(object viewModel, object ownerViewModel, string? context = null);
+
+    /// <summary>
+    /// Displays an open file picker dialog.
+    /// </summary>
+    /// <param name="options">The dialog options.</param>
+    /// <param name="ownerViewModel">The owner view model.</param>
+    /// <returns>A list of selected files.</returns>
+    Task<IReadOnlyList<PickFileResult>> ShowFileOpenPickerAsync(FileOpenPickerOptions options, object ownerViewModel);
+
+    /// <summary>
+    /// Displays a save file picker dialog.
+    /// </summary>
+    /// <param name="options">The dialog options.</param>
+    /// <param name="ownerViewModel">The owner view model.</param>
+    /// <returns>The selected file</returns>
+    Task<PickFileResult?> ShowFileSavePickerAsync(FileSavePickerOptions options, object ownerViewModel);
+
+    /// <summary>
+    /// Displays an open folder picker dialog.
+    /// </summary>
+    /// <param name="options">The dialog options.</param>
+    /// <param name="ownerViewModel">The owner view model.</param>
+    /// <returns>The selected folder.</returns>
+    Task<PickFolderResult?> ShowFolderPickerAsync(FolderPickerOptions options, object ownerViewModel);
 }