+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Added
 - Avalonia: Added file-dialogs to `IWindowManager`
+- WinUI: Added file-dialogs to `IWindowManager`
 
 ## [6.1.0] - 2025-12-08